graph_auth.py
# graph_auth.py
import os
import requests
from dotenv import load_dotenv

load_dotenv()

TENANT_ID = os.getenv("TENANT_ID")
CLIENT_ID = os.getenv("CLIENT_ID")
CLIENT_SECRET = os.getenv("CLIENT_SECRET")

TOKEN_URL = f"https://login.microsoftonline.com/{TENANT_ID}/oauth2/v2.0/token"
SCOPE = "https://graph.microsoft.com/.default"


def get_graph_token():
    """
    Returns an application-only Microsoft Graph access token.
    Raises if authentication fails.
    """
    data = {
        "client_id": CLIENT_ID,
        "client_secret": CLIENT_SECRET,
        "grant_type": "client_credentials",
        "scope": SCOPE,
    }

    resp = requests.post(TOKEN_URL, data=data)
    resp.raise_for_status()

    return resp.json()["access_token"]
