requirements.txt
absl-py==2.3.1
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.2
idna==3.11
immutabledict==4.2.2
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.0.2
openpyxl==3.1.5
ortools==9.15.6755
pandas==2.3.3
protobuf>=6.33.1,<6.34
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
Werkzeug==3.1.3
zipp==3.23.0
